pyproject.toml
[tool.poetry]
name = "machine-learning"
version = "0.1.0"
description = ""
authors = ["nitichaisawangsai"]
license = "nitichaisawangsai"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.2.2"
datetime = "^5.5"
logging = "^0.4.9.6"
pytz = "^2024.1"
numpy = "^1.26.4"
lab = "^8.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
